Limit the number of jenkins slave pods a user can run at a time to 1
This is a sample of containers active while we run a maven project like the
vertx quickstart.

| image                                         | name  | CPU   | Memory |
|                                               |       |       |        |
| 'fabric8/maven-builder:v9ff62e0'              | maven | 1250m | 640Mi  |
| 'fabric8/jenkins-slave-base-centos7:vb0268ae' | jnlp  | 500m  | 256Mi  |

The total 896Mi and 1750m CPU implies only one of these pods can be scheduled at
a time. As far as I understand, we are wasting the rest of (1G - 896M) memory
and (2 - 1.750) CPU.