Add content-repository boot dependency to Jenkins
This commit adds a hard dependency on the Jenkins deployment such
that Jenkins will not be brought up unless content-repository is
not up and running.

This is done by adding init containers to the Jenkins' Kubernetes
Deployment files which check every 1 second if content-repository
is up or not. The check is set to timeout at 4 seconds. Once the
check returns a 0 exit code and the init container exits, only
then the Jenkins pod is brought up.