Lower retry count in Jenkins init containers
This commit lowers the retry count in Jenkins init containers from
100 to 10. This is done because due to an underlying OpenShift
networking issue at
https://github.com/openshiftio/openshift.io/issues/3299, the init
container is not always able to talk to the content-repository
reliably leading to failure in bringing up Jenkins. With a low
retry count i.e. 10, the init container will fail fast if this
intermittent network failure occurs, and another pod will be
started by OpenShift which should be able to talk to
content-repository, hopefully!