Adapt to change to mavenCI step
Recently a change was introduced to mavenCI step which runs `install`
goal instead of `deploy`, this breaks testing PR as artifacts won't be
uploaded to nexus.

This patch fixes the issue by overriding the goal to `deploy`.

For details about the changes to mavenCI and mavenCanaryRelease
See: https://github.com/fabric8io/fabric8-pipeline-library/pull/418